backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
click==8.1.8
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.2
PyMySQL==1.1.1
pytest==8.4.1
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
sqlparse==0.5.3
starlette==0.46.1
typing_extensions==4.12.2
tzdata==2025.2
uvicorn==0.34.0
